webapi/tct-deviceorientation-w3c-tests/deviceorientation/DeviceMotionEvent_attributes_test.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Xinx, liu <xinx.liu@intel.com>

-->

<meta charset='utf-8'>
<title>Orientation Test: DeviceMotionEvent attributes exists and type are right</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://dev.w3.org/geo/api/spec-source-orientation.html#devicemotion">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
  setup({explicit_done: true, timeout: 2000});
  var run = false;
  on_event(window, "devicemotion", function(event) {
    if(!run) {
      run = true;
      test(function() {
        assert_true("acceleration" in event, "Attribute acceleration missing in DeviceMotionEvent object");
      }, "Check if the DeviceMotionEvent.acceleration method exists");

      test(function() {
        assert_equals(typeof event.acceleration, "object", "Attribute acceleration in DeviceMotionEvent of type");
      }, "Check if DeviceMotionEvent.acceleration is type of DeviceAcceleration");

      test(function() {
        assert_true("accelerationIncludingGravity" in event, "Attribute accelerationIncludingGravity missing in DeviceMotionEvent object");
      }, "Check if the DeviceMotionEvent.accelerationincludinggravity method exists");

      test(function() {
        assert_equals(typeof event.accelerationIncludingGravity, "object", "Attribute accelerationIncludingGravity in DeviceMotionEvent of type");
      }, "Check if DeviceMotionEvent.accelerationIncludingGravity is type of DeviceAcceleration");

      test(function() {
        assert_true("interval" in event, "Attribute interval missing in DeviceMotionEvent object");
      }, "Check if the DeviceMotionEvent.interval attribute exists");

      test(function() {
        assert_true("rotationRate" in event, "Attribute rotationRate missing in DeviceMotionEvent object");
      }, "Check if the DeviceMotionEvent.rotationRate attribute exists");

      test(function() {
        assert_equals(typeof event.rotationRate, "object", "Attribute rotationRate in DeviceMotionEvent of type");
      }, "Check if DeviceMotionEvent.rotationRate is type of DeviceRotationRate");

      done();

      if(t) {
         clearTimeout(t);
      }
    }
  });
  var t = setTimeout(function () {
    test(function() {
      assert_unreached("Can not fire devicemotion event");
    }, "Check if the DeviceMotionEvent.acceleration method exists");
    test(function() {
      assert_unreached("Can not fire devicemotion event");
    }, "Check if DeviceMotionEvent.acceleration is type of DeviceAcceleration");
    test(function() {
      assert_unreached("Can not fire devicemotion event");
    }, "Check if the DeviceMotionEvent.accelerationincludinggravity method exists");
    test(function() {
      assert_unreached("Can not fire devicemotion event");
    }, "Check if DeviceMotionEvent.accelerationIncludingGravity is type of DeviceAcceleration");
    test(function() {
      assert_unreached("Can not fire devicemotion event");
    }, "Check if the DeviceMotionEvent.interval attribute exists");
    test(function() {
      assert_unreached("Can not fire devicemotion event");
    }, "Check if the DeviceMotionEvent.rotationRate attribute exists");
    test(function() {
      assert_unreached("Can not fire devicemotion event");
    }, "Check if DeviceMotionEvent.rotationRate is type of DeviceRotationRate");
    done();
  }, 1500)
</script>
